restructure mmu code
