+++ src/internals/mmu/mod.rs
@@ -1,0 +1,5 @@
+pub mod l1;
+pub mod l2;
+pub mod setup;
+
+pub use setup::initialize;
