+++ src/internals/tasks.rs
@@ -175,7 +175,6 @@ impl Scheduler {
     }
 }
 
-#[link_section = "user_stack_start"]
 static mut SCHEDULER: Scheduler = Scheduler::new();
 
 #[allow(static_mut_refs)]
